Create scheduler state module
Moving enough code out of `simple_scheduler` for decoupling stateful
structures to support trait based interfaces in the future. There is
no logical flow change in this refactor, all tests pass. Some
structures will have `pub(crate)` visable modules and fields to help
with the transition of code reordering.
