+++ nativelink-scheduler/BUILD.bazel
@@ -17,6 +17,11 @@ rust_library(
         "src/operation_state_manager.rs",
         "src/platform_property_manager.rs",
         "src/property_modifier_scheduler.rs",
+        "src/scheduler_state/awaited_action.rs",
+        "src/scheduler_state/completed_action.rs",
+        "src/scheduler_state/mod.rs",
+        "src/scheduler_state/state_manager.rs",
+        "src/scheduler_state/workers.rs",
         "src/simple_scheduler.rs",
         "src/worker.rs",
         "src/worker_scheduler.rs",
