+++ nativelink-scheduler/src/scheduler_state/state_manager.rs
@@ -1,0 +1,66 @@
+// Copyright 2024 The NativeLink Authors. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//    http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use std::collections::BTreeMap;
+use std::sync::Arc;
+
+use hashbrown::{HashMap, HashSet};
+use nativelink_util::action_messages::ActionInfo;
+
+use crate::scheduler_state::awaited_action::AwaitedAction;
+use crate::scheduler_state::completed_action::CompletedAction;
+use crate::scheduler_state::workers::Workers;
+
+/// StateManager is responsible for maintaining the state of the scheduler. Scheduler state
+/// includes the actions that are queued, active, and recently completed. It also includes the
+/// workers that are available to execute actions based on allocation strategy.
+pub(crate) struct StateManager {
+    // TODO(adams): Move `queued_actions_set` and `queued_actions` into a single struct that
+    //  provides a unified interface for interacting with the two containers.
+
+    // Important: `queued_actions_set` and `queued_actions` are two containers that provide
+    // different search and sort capabilities. We are using the two different containers to
+    // optimize different use cases. `HashSet` is used to look up actions in O(1) time. The
+    // `BTreeMap` is used to sort actions in O(log n) time based on priority and timestamp.
+    // These two fields must be kept in-sync, so if you modify one, you likely need to modify the
+    // other.
+    /// A `HashSet` of all actions that are queued. A hashset is used to find actions that are queued
+    /// in O(1) time. This set allows us to find and join on new actions onto already existing
+    /// (or queued) actions where insert timestamp of queued actions is not known. Using an
+    /// additional `HashSet` will prevent us from having to iterate the `BTreeMap` to find actions.
+    ///
+    /// Important: `queued_actions_set` and `queued_actions` must be kept in sync.
+    pub(crate) queued_actions_set: HashSet<Arc<ActionInfo>>,
+
+    /// A BTreeMap of sorted actions that are primarily based on priority and insert timestamp.
+    /// `ActionInfo` implements `Ord` that defines the `cmp` function for order. Using a BTreeMap
+    /// gives us to sorted actions that are queued in O(log n) time.
+    ///
+    /// Important: `queued_actions_set` and `queued_actions` must be kept in sync.
+    pub(crate) queued_actions: BTreeMap<Arc<ActionInfo>, AwaitedAction>,
+
+    /// A `Workers` pool that contains all workers that are available to execute actions in a priority
+    /// order based on the allocation strategy.
+    pub(crate) workers: Workers,
+
+    /// A map of all actions that are active. A hashmap is used to find actions that are active in
+    /// O(1) time. The key is the `ActionInfo` struct. The value is the `AwaitedAction` struct.
+    pub(crate) active_actions: HashMap<Arc<ActionInfo>, AwaitedAction>,
+
+    /// These actions completed recently but had no listener, they might have
+    /// completed while the caller was thinking about calling wait_execution, so
+    /// keep their completion state around for a while to send back.
+    /// TODO(#192) Revisit if this is the best way to handle recently completed actions.
+    pub(crate) recently_completed_actions: HashSet<CompletedAction>,
+}
